atomic-agent-azure-elastic-observability/agent/requirements.txt
elastic-apm
elasticsearch
python-dotenv
requests
PyYAML
